refactor: 세션ID 생성과 저장하는 기능을 한 곳에 합침
